fix: compatible type

+++ src/main/java/com/example/eduworldbe/dto/GradeExamResponse.java
@@ -8,6 +8,6 @@ public class GradeExamResponse {
   private String attemptId;
   private String examId;
   private String userId;
-  private Integer totalScore; // Tổng điểm
-  private Map<String, Integer> questionScores; // Map chứa điểm từng câu: key là questionId, value là điểm
+  private Double totalScore; // Tổng điểm
+  private Map<String, Double> questionScores; // Map chứa điểm từng câu: key là questionId, value là điểm
 }
